+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Release Notes
 ## [Unreleased](https://github.com/ytekeli/dom-selector/compare/v0.1.0...master)
 
+### Added
+- Add `extractFromFile` and `extractFromUrl` features.
+- Improve some test cases.
+
 ## [v0.1.0 (2021-06-25)](https://github.com/ytekeli/dom-selector/compare/v0.1.0...master)
 
 First version is released.